리팩토링 : search 상태 내부에서 관리
